requirements.txt
Pillow
google-cloud-vision
google-cloud-language
fastapi
uvicorn
pydantic
requests
python-dateutil
python-multipart
httpx